requirements.txt
django
djangorestframework
gunicorn
openai>=1.0.0
scikit-learn
numpy
umap-learn
plotly